+++ api/db.php
@@ -1,5 +1,5 @@
 <?php
-    //Connecto to db
+    //Connect to db
     $uri = 'mongodb://localhost:27017';
     $client = new MongoDB\Client($uri);
