+++ index.html
@@ -2,10 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" href="./public/logo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./src/bootstrap.min.css">
-    <title>Vite + React</title>
+    <title>Stocks app</title>
   </head>
   <body>
     <div id="root"></div>
