+++ src/components/Login.jsx
@@ -18,7 +18,6 @@ export function Login () {
         }
 
         login(newUser).then((response) => {
-            console.log(response)
             if(response.status == 200) {
                 localStorage.setItem('token', response.data.token)
                 navigate('/')
